pyproject.toml
[tool.poetry]
name = "production-control-system"
version = "0.1.0"
description = "The functionality of the system is to receive shift tasks (batch) and unique product identifiers within this batch, as well as check (upon request from an external system) whether a given product identifier belongs to this batch."
authors = ["ShadeSWD <seawolfy60@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^69.0.3"
uvicorn = "^0.27.0.post1"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
